[URGENT] Fix to most commands and celldoors (#7)
* Create jbulx-showdeath.lua

* Added "make warden" command and "teleport guards

* Added "make warden" & "teleport guards" commands

* Create jbulx-celldoors.lua

* Update jbulx-showdeath.lua

* Update README.md

* Differentiate revive and respawn

Changed old "revive" command to "respawn" and added a new revive command that only respawns the target if they are not already alive or in spectator mode.

* Fix revive announcement, add slaynr commands

Fixed revive logging to only include those actually being revived, rather than all intended targets.
Added commands "slaynr", "rslaynr" and "cslaynr" (an edited version of the same commands for TTT).
Added hook for round start for slaynr.

* Updated readme

Added slaynr commands and revive/respawn differentiation to list of commands.
Updated copyright year to 2018.

* Fixed GNU licence section of readme

*Now has hyperlink to website.
*"License" changed to "licence" (license is a verb, licence is a noun).

* Removed "hyperlink"

* Update jbulx-teams.lua

* Add delay to tsay

Added a tick delay to the tsays that didn't have them already.

* Fix small issue in command list

ulx cslaynr did not have the <player> tag it should have had. It does now.

* Added wardenban

*Added ulx.wardenban
*Added ulx.wardenbaninfo
*Added hook for detecting player claiming warden
*Added new recommendation and default for use of guard/wardenban info

* Major additions, changes and merging

*Added ulx helicopter (created by Owen)
*Added ulx mancannon (created by Owen)
*Redid celldoors command to allow doors to be toggled open and closed (this does not work on carceris, so the command will only open doors on that map)
*Added support for some maps on celldoors
*Merged ulx helicopter, celldoors and mancannon into one file: jbulx-mapinteract.lua
*Merged jbulx-showdeath.lua into jbulx-misc.lua
*Added wardenban commands
*Merged rslaynr into slaynr through setOpposite
*Shortened wrong gamemode error into one line
*Added checking for makeguard, makeprisoner and makespectator to prevent forcing players to a team they are already on
*Changed makewarden to only allow one target
*Redid teleport guards to be more efficient and work better
*Added more chat command aliases for most commands
*Other, smaller changes
*Updated README.md to show changes

Co-Authored-By: Owen <pn-owen@users.noreply.github.com>

* Fixes and return to doors open

Fixed jailbreak gamemode detection - used to do the opposite of what it was supposed to.
Returned togglecells to open, not toggle, now known as opencells. This was done as certain maps (e.g. vipinthemix and carceris) and it wasn't very useful to close cells using a command anyway.

* Function name fix and readme update